Moved the classes to the correct place and added the variables
